made trasmission_grad to a pytest code
